printing polys over big nums
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk/M2@6147 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/d/parser.d
@@ -27,7 +27,7 @@ export parseRR(s:string):RR := (			    -- 4.33234234234p345e-9
      extra := 16;
      prec := 53;
      exsign := false;
-     expon := 0;
+     expon := long(0);
      foreach c in s do (
 	  if c == 'e' || c == 'E' then (
 		    inPrec = false;
